separando actiosn de reducers
