+++ src/components/molecules/NavBar/style.js
@@ -6,7 +6,7 @@ export const PageNavbar = styled.div`
     box-shadow: rgba(0, 0, 0, 0.15) 0px 5px 10px 0px;
     justify-content: space-between;
     align-items: center;
-    padding: 0.3em 2em;
+    padding: 0.3em 0.5em;
 `
 export const StyleLink = styled(Link)`
     text-decoration: none;
@@ -15,9 +15,9 @@ export const StyleLink = styled(Link)`
     transition: 0.7s;
     font-size: 1em;
     font-family: "Montserrat", sans-seri;
+    margin: -0.1em -1.8em;
+    margin-right: 1em;
     
-    
-
 
 &:hover{
     color: #B0C4DE;
@@ -31,7 +31,7 @@ export const LinkList = styled.ul`
 
 `
 export const LinkListLi = styled.li`
-    margin-right: 3em;
+    margin-right: 1em;
     padding: 1em;
 
 `
@@ -42,7 +42,6 @@ export const CardIcon = styled.div`
     justify-content: flex-end;
     align-items: flex-end;
     position: absolute;
-    margin: 0 20px;
     
 svg {
     cursor: pointer;
