+++ src/components/pages/login/logout.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { ButtonLogin } from '../../atoms/Bottons/ButtonLogin/style'
 import {useDispatch} from 'react-redux'
-import { logout } from '../../../redux/reducers/userSlice'
+
 import { Navigate } from 'react-router'
 import { StyleLink } from '../../molecules/NavBar/style'
 
@@ -9,7 +9,7 @@ const Logout = () => {
     const dispatch = useDispatch()
 
     const handleLogout = () =>{
-        dispatch(logout())
+  
         
     }
   return (
