+++ src/redux/actions/Users.js
@@ -1,0 +1,8 @@
+import *as types from './types'
+
+export function usersLogin(username, password){
+    return{
+        type: types.USERS,
+        payload: { username, password }
+    }
+}
