+++ src/redux/actions/types.js
@@ -1,0 +1,1 @@
+export const USERS = 'USERS';
