+++ src/redux/rootReducer.js
@@ -1,0 +1,8 @@
+import  {combineReducers} from 'redux'
+
+import modelsReducer from './reducers/reducer'
+
+export default combineReducers({
+    usuario:modelsReducer,
+
+})
